src/backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=1.4
python-jose[crytography]
passlib[bcrypt]
python-dotenv
pydantic[email]
pytest
python-multipart
httpx
